Make DefaultSessionHandler and GlobalSessionHandler into inner classes of SessionHandler
